Rename ShardBlockBloomFilter to ManagedBloomFilter
