+++ persist/fs/seek.go
@@ -87,7 +87,7 @@ type seeker struct {
 
 	// Bloom filter associated with the shard / block the seeker is responsible
 	// for. Needs to be closed when done.
-	bloomFilter block.ShardBlockBloomFilter
+	bloomFilter block.ManagedBloomFilter
 }
 
 type indexMapEntry struct {
