+++ persist/fs/types.go
@@ -68,7 +68,7 @@ type FileSetReader interface {
 	ReadMetadata() (id ts.ID, length int, checksum uint32, err error)
 
 	// ReadBloomFilter returns the bloom filter stored on disk
-	ReadBloomFilter() (block.ShardBlockBloomFilter, error)
+	ReadBloomFilter() (block.ManagedBloomFilter, error)
 
 	// Validate validates the data and returns an error if the data are corrupted
 	Validate() error
